refactor: search and scan to the backend component
